move NotificationDetailRouteState in user model; remove NotificationData export
